Server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.61.0
PyMuPDF==1.25.2
Pillow==11.1.0
reportlab==4.2.5
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.1.7
email-validator>=2.0.0
